requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
db-dtypes==1.2.0
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
httplib2==0.22.0
idna==3.7
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
